Improve search
Search is able to, to some degree of accuracy, determine artist and
song, and uses that to decide weather or not to fetch new songs online
or play an existing one.

When finding multiple hits of any song online, I now attempt to remove
duplicates. The rules for which songs have a higher priority are now
deterministic and favors shorter artist names and exact matches of
names (word by word).