Improve background downloading while listening
Google searches have a timeout threshold which if breached will
disallow searches for some time. When using --background-download
while listening, the player won't stall for waiting the for
threshold, but instead move on to the next song.

Also fixed various search bugs (such as Youtube /channel/ URLs)
and keeping on playing when in --background-download if URL missing.